[BUGFIX] Fixed say command using slash commands as it would only work with prefix.

+++ cogs/say.py
@@ -21,9 +21,11 @@ class SayCommandCog(commands.Cog):
     @commands.hybrid_command(description="Use this command to say things to people using the bot.", with_app_command=True, extras={"category": "General"})
     async def say(self, ctx, *, message: str):
         
-
+        
         try:
-            await ctx.message.delete()
+            if ctx.message:
+                await ctx.message.delete()
+                
             await ctx.send(message)
         except Exception as e:
             error_id = shortuuid.ShortUUID().random(length=8)
